// src/lib/api-client/models/bulk-enriched-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Studio API
 * Studio API for managing campaigns, personas, proxies, and keyword sets
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BulkMetadata } from './bulk-metadata';
// May contain unused imports in some cases
// @ts-ignore
import type { EnrichedCampaignData } from './enriched-campaign-data';

/**
 * 
 * @export
 * @interface BulkEnrichedDataResponse
 */
export interface BulkEnrichedDataResponse {
    /**
     * 
     * @type {{ [key: string]: EnrichedCampaignData; }}
     * @memberof BulkEnrichedDataResponse
     */
    'campaigns'?: { [key: string]: EnrichedCampaignData; };
    /**
     * 
     * @type {BulkMetadata}
     * @memberof BulkEnrichedDataResponse
     */
    'metadata'?: BulkMetadata;
    /**
     * 
     * @type {number}
     * @memberof BulkEnrichedDataResponse
     */
    'totalCount'?: number;
}